// src/app/model/usuario.dto.ts
export interface UsuarioNuevodto {
  name: string;
  lastname: string;
  username: string;
  password: string;
}
export interface UsuarioActualizadodto {
  id: number;
  name: string;
  lastname: string;
  username: string;
  password: string;
  created: Date;
}
export interface Usuariodto {
  id: number;
  name: string;
  lastname: string;
  username: string;
  password: string;
}